+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- next-header -->
 ## [Unreleased]
 
+## [0.17.4] - 2025-07-12
+
 ### Added
 
 - Removed `#[cfg(feature = "alloc")]` bound on `BumpBox::into_box`
@@ -701,7 +703,8 @@ _If you are upgrading: please see [`UPGRADING.md#0.17.0`](UPGRADING.md#0.17.0)._
 [#3]: https://github.com/bluurryy/bump-scope/issues/3
 
 <!-- next-url -->
-[Unreleased]: https://github.com/bluurryy/bump-scope/compare/v0.17.3...HEAD
+[Unreleased]: https://github.com/bluurryy/cargo-insert-docs/compare/v0.17.4...HEAD
+[0.17.4]: https://github.com/bluurryy/bump-scope/releases/tag/v0.17.4
 [0.17.3]: https://github.com/bluurryy/bump-scope/releases/tag/v0.17.3
 [0.17.2]: https://github.com/bluurryy/bump-scope/releases/tag/v0.17.2
 [0.17.1]: https://github.com/bluurryy/bump-scope/releases/tag/v0.17.1